# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Beer Catalogue API
  description: API for managing a catalogue of beers.
  version: 0.0.1

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Beers
    description: Beer catalogue operations
  - name: Manufacturers
    description: Beer manufacturer operations

paths:
  /api/beers:
    get:
      tags: [Beers]
      operationId: getBeers
      summary: Get all beers
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [name, abv, type]
            default: name
        - name: sortDirection
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: search
          in: query
          required: false
          schema:
            type: string
            description: Case-insensitive search across name, type, and manufacturer name
      responses:
        '200':
          description: A paginated list of beers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeerPage'

    post:
      tags: [Beers]
      operationId: createBeer
      summary: Add a new beer to the catalogue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeerCreateRequest'
      responses:
        '201':
          description: Beer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beer'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Invalid beer type
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /api/beers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Beers]
      operationId: getBeer
      summary: Get a beer by ID
      responses:
        '200':
          description: Beer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beer'
        '404':
          description: Beer not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

    put:
      tags: [Beers]
      operationId: updateBeer
      summary: Update an existing beer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeerUpdateRequest'
      responses:
        '200':
          description: Beer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beer'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Beer not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Invalid beer type
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

    delete:
      tags: [Beers]
      operationId: deleteBeer
      summary: Delete a beer
      responses:
        '204':
          description: Beer deleted
        '404':
          description: Beer not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /api/manufacturers:
    get:
      tags: [Manufacturers]
      operationId: getManufacturers
      summary: Get all manufacturers
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: A paginated list of manufacturers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManufacturerPage'

    post:
      tags: [Manufacturers]
      operationId: createManufacturer
      summary: Add a new manufacturer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturerCreateRequest'
      responses:
        '201':
          description: Manufacturer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Invalid country code
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /api/manufacturers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Manufacturers]
      operationId: getManufacturer
      summary: Get a manufacturer by ID
      responses:
        '200':
          description: Manufacturer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
        '404':
          description: Manufacturer not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

    put:
      tags: [Manufacturers]
      operationId: updateManufacturer
      summary: Update an existing manufacturer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturerUpdateRequest'
      responses:
        '200':
          description: Manufacturer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Manufacturer not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Invalid country code
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

    delete:
      tags: [Manufacturers]
      operationId: deleteManufacturer
      summary: Delete a manufacturer
      responses:
        '204':
          description: Manufacturer deleted
        '404':
          description: Manufacturer not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

components:
  schemas:
    Beer:
      type: object
      required: [id, name, type, abv, manufacturer]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          example: Zipfer Urquell
        type:
          type: string
          x-extensible-enum:
            - LAGER
            - PILSNER
            - HELLES
            - MAERZEN
            - DUNKEL
            - BOCK
            - DOPPELBOCK
            - PALE_ALE
            - IPA
            - DOUBLE_IPA
            - SESSION_IPA
            - AMBER_ALE
            - BROWN_ALE
            - PORTER
            - STOUT
            - IMPERIAL_STOUT
            - BARLEYWINE
            - HEFEWEIZEN
            - WITBIER
            - WHEAT
            - BERLINER_WEISSE
            - SAISON
            - FARMHOUSE_ALE
            - DUBBEL
            - TRIPEL
            - QUADRUPEL
            - SOUR
            - GOSE
            - LAMBIC
            - CREAM_ALE
            - KOLSCH
            - ALTBIER
          example: LAGER
        abv:
          type: number
          format: double
          description: Alcohol by volume in percent
          example: 4.8
        description:
          type: string
          example: A crisp and refreshing lager.
        manufacturer:
          $ref: '#/components/schemas/Manufacturer'
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true

    Manufacturer:
      type: object
      required: [id, name, originCountry]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        originCountry:
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true

    BeerCreateRequest:
      type: object
      required: [name, type, abv, manufacturerId]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        type:
          type: string
          minLength: 1
          maxLength: 50
        abv:
          type: number
          format: double
          minimum: 0.0
          maximum: 100.0
        description:
          type: string
          maxLength: 500
        manufacturerId:
          type: string
          format: uuid

    BeerUpdateRequest:
      type: object
      required: [name, type, abv, description, manufacturerId]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        type:
          type: string
          minLength: 1
          maxLength: 50
        abv:
          type: number
          format: double
          minimum: 0.0
          maximum: 100.0
        description:
          type: string
          maxLength: 500
        manufacturerId:
          type: string
          format: uuid

    ManufacturerCreateRequest:
      type: object
      required: [name, originCountry]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        originCountry:
          type: string
          minLength: 1
          maxLength: 100

    ManufacturerUpdateRequest:
      type: object
      required: [name, originCountry]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        originCountry:
          type: string
          minLength: 1
          maxLength: 100

    BeerPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Beer'
        page:
          type: integer
        size:
          type: integer
        hasMore:
          type: boolean

    ManufacturerPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Manufacturer'
        page:
          type: integer
        size:
          type: integer
        hasMore:
          type: boolean

    Problem:
      type: object
      description: application/problem+json conform error response
      properties:
        title:
          type: string
          description: Short, stable summary of the problem type
          example: Beer not found
        status:
          type: integer
          description: HTTP status code
          example: 404
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: No beer with id 'a3b8c2d1-...' exists in the catalogue
        instance:
          type: string
          format: uri
          description: URI identifying this specific problem occurrence
          example: /api/beers/a3b8c2d1-